requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
pydantic[email]==2.5.0
google-generativeai==0.8.5
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
librosa==0.10.1
numpy==1.24.3